backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
qdrant-client>=1.9.0
redis>=5.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pypdf>=3.17.0
sentence-transformers>=2.2.0
python-multipart>=0.0.6
